# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.AccelerationSensor:
    properties:
      measurements:
        items: {}
        type: array
      type:
        type: string
    type: object
  model.Gateway:
    properties:
      id:
        type: string
      ip_address:
        type: string
      last_contact:
        type: string
      network_segment:
        type: integer
      online:
        type: boolean
      tags:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
    type: object
  model.GatewayConfig:
    properties:
      api_timeout:
        type: number
      max_allowed_clients:
        type: integer
      poll_interval:
        type: integer
      version:
        type: string
    type: object
  model.Tag:
    properties:
      address:
        type: string
      last_contact:
        type: string
      name:
        type: string
      online:
        type: boolean
      sensors:
        items: {}
        type: array
    type: object
info:
  contact: {}
paths:
  /acc-data/get/{tag}/{page}:
    get:
      consumes:
      - application/json
      description: gets data of the specified tag - get your tags via "list" first
      parameters:
      - description: the tags address
        example: '"12:34:56:78:90:12"'
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: gets acc data of the specified tag
      tags:
      - acc-data
  /acc-data/list:
    get:
      consumes:
      - application/json
      description: gets data of the specified tag - get your tags via "list" first
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.AccelerationSensor'
            type: array
      summary: gets acc data of the specified tag
      tags:
      - acc-data
  /config/get/{gatewayid}:
    get:
      consumes:
      - application/json
      description: get config of a specific gateway to be able to decide on version
        updates etc. To find the gateways id use /structure/gateway/list
      parameters:
      - description: the gateways id
        in: path
        name: gatewayid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GatewayConfig'
            type: array
      summary: lists available configs
      tags:
      - config
  /config/get/{gatewayid}/{tagaddress}:
    get:
      consumes:
      - application/json
      description: get config of a specific gateway to be able to decide on version
        updates etc. To find the gateways id use /structure/gateway/list, to get the
        tags address use /structure/tag/list/{gatewayid}
      parameters:
      - description: the gateways id
        in: path
        name: gatewayid
        required: true
        type: string
      - description: the tags id
        in: path
        name: tagaddress
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GatewayConfig'
            type: array
      summary: lists available configs
      tags:
      - config
  /ping:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - ping
  /structure/gateway/list:
    get:
      consumes:
      - application/json
      description: lists gateways
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Gateway'
            type: array
      summary: lists available gateways
      tags:
      - structure
  /structure/tag/list/{gatewayid}:
    get:
      consumes:
      - application/json
      description: lists tags
      parameters:
      - description: id of the gateway
        in: path
        name: gateway_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Gateway'
            type: array
      summary: lists available tags - get the gateway_id from structure/gateway/list
      tags:
      - structure
swagger: "2.0"